requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-Web-Log==0.1
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.4
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.31.0
requests-toolbelt==1.0.0
responses==0.25.0
simple-websocket==1.1.0
SQLAlchemy==2.0.38
sqlparse==0.4.4
stripe==11.6.0
urllib3==2.2.1
Werkzeug==3.1.3
WTForms==3.2.1